Revert "SUBS-472_dispatching-validation-messages"
