+++ subs-api/src/main/java/uk/ac/ebi/subs/api/handlers/CoreSubmittableEventHelper.java
@@ -5,8 +5,8 @@ import org.springframework.data.rest.core.annotation.HandleBeforeCreate;
 import org.springframework.data.rest.core.annotation.HandleBeforeSave;
 import org.springframework.data.rest.core.annotation.RepositoryEventHandler;
 import org.springframework.stereotype.Component;
-import uk.ac.ebi.subs.api.services.SubmittableValidationDispatcher;
-import uk.ac.ebi.subs.repository.model.*;
+import uk.ac.ebi.subs.repository.model.ProcessingStatus;
+import uk.ac.ebi.subs.repository.model.StoredSubmittable;
 import uk.ac.ebi.subs.repository.repos.status.ProcessingStatusRepository;
 
 import java.util.UUID;
@@ -16,13 +16,11 @@ import java.util.UUID;
 public class CoreSubmittableEventHelper {
 
     @Autowired
-    public CoreSubmittableEventHelper(ProcessingStatusRepository processingStatusRepository, SubmittableValidationDispatcher submittableValidationDispatcher) {
+    public CoreSubmittableEventHelper(ProcessingStatusRepository processingStatusRepository) {
         this.processingStatusRepository = processingStatusRepository;
-        this.submittableValidationDispatcher = submittableValidationDispatcher;
     }
 
     private ProcessingStatusRepository processingStatusRepository;
-    private SubmittableValidationDispatcher submittableValidationDispatcher;
 
 
     /**
@@ -31,7 +29,7 @@ public class CoreSubmittableEventHelper {
      * @param submittable
      */
     @HandleBeforeCreate
-    public void setProcessingStatus(StoredSubmittable submittable) {
+    public void beforeCreate(StoredSubmittable submittable) {
         submittable.setId(UUID.randomUUID().toString());
 
         ProcessingStatus processingStatus = ProcessingStatus.createForSubmittable(submittable);
@@ -40,32 +38,6 @@ public class CoreSubmittableEventHelper {
 
         setTeamFromSubmission(submittable);
     }
-
-    // Validation of created submittables
-    @HandleBeforeCreate
-    public void validateOnCreate(Sample sample) { submittableValidationDispatcher.validateCreate(sample); }
-
-    @HandleBeforeCreate
-    public void validateOnCreate(Study study) { submittableValidationDispatcher.validateCreate(study); }
-
-    @HandleBeforeCreate
-    public void validateOnCreate(Assay assay) {
-        submittableValidationDispatcher.validateCreate(assay);
-    }
-
-    // Validation of updated submittables
-
-    @HandleBeforeSave
-    public void validateOnSave(Sample sample) { submittableValidationDispatcher.validateUpdate(sample); }
-
-    @HandleBeforeSave
-    public void validateOnSave(Study study) { submittableValidationDispatcher.validateUpdate(study); }
-
-    @HandleBeforeSave
-    public void validateOnSave(Assay assay) {
-        submittableValidationDispatcher.validateUpdate(assay);
-    }
-
 
     private void setTeamFromSubmission(StoredSubmittable submittable) {
         if (submittable.getSubmission() != null) {
