+++ subs-api/src/main/java/uk/ac/ebi/subs/api/services/SubmissionEventServiceImpl.java
@@ -29,7 +29,6 @@ public class SubmissionEventServiceImpl implements SubmissionEventService {
     @Override
     public void submissionCreated(Submission submission) {
         //TODO
-
     }
 
     @Override
