+++ subs-api/src/test/java/uk/ac/ebi/subs/api/ApiIntegrationTest.java
@@ -49,6 +49,8 @@ public class ApiIntegrationTest {
     private int port;
     private String rootUri;
 
+
+
     private ApiIntegrationTestHelper testHelper;
 
     @Autowired
@@ -71,6 +73,24 @@ public class ApiIntegrationTest {
         submissionRepository.deleteAll();
         sampleRepository.deleteAll();
         submissionStatusRepository.deleteAll();
+
+        Unirest.setObjectMapper(new com.mashape.unirest.http.ObjectMapper() {
+            public <T> T readValue(String value, Class<T> valueType) {
+                try {
+                    return objectMapper.readValue(value, valueType);
+                } catch (IOException e) {
+                    throw new RuntimeException(e);
+                }
+            }
+
+            public String writeValue(Object value) {
+                try {
+                    return objectMapper.writeValueAsString(value);
+                } catch (JsonProcessingException e) {
+                    throw new RuntimeException(e);
+                }
+            }
+        });
 
         testHelper = new ApiIntegrationTestHelper(objectMapper,rootUri);
     }
