+++ subs-data-model/src/main/java/uk/ac/ebi/subs/data/submittable/BaseSubmittable.java
@@ -120,5 +120,4 @@ public abstract class BaseSubmittable<T extends BaseSubmittable> implements Subm
     public void setAttributes(List<Attribute> attributes) {
         this.attributes = attributes;
     }
-
 }
