+++ subs-samples-agent/src/test/resources/application.properties
@@ -1,2 +1,2 @@
 biosamples.client.uri=http://snowy.ebi.ac.uk:8081/biosamples/beta/
-test.accession=SAMEA100001
+test.accession=SAMEA100002
